working version, can't handle overflow

+++ include/GaussianBlur.h
@@ -15,7 +15,7 @@ class Gaussian_Blur {
     std::vector<std::vector<int>> variances;
     int variance_to_depth(float var);
 public:
-    void generate_binomial_distrib(int n, std::vector<float> & new_distrib);
+    void generate_binomial_distrib(int n, std::vector<int> & new_distrib);
     int generate_kernel(std::vector<int> & kernel, float var);
     void convolve(Image & img, float var);
 };
